fix: Custom Resources export were broken on exported games
We don't have access to the gdscript on exported game where the code
base is compiled and not text, so this stops working. I had to save the
variables found in the editor to the meta so that when looking up the
source code is not available, you can just use the saved meta.

But this code is still kind of messy, I'd like to refactor it some day