Refactor to stop using meta o editor related utilities
meta is not saved in inherited scenes, so it is not reliable
and broke muto project. Now I'm usng the original node as a
"backing field" when using only CustomInspector, which is already
pretty good and I didn't need to use meta from the start.

for CustomEditoControls I'm using a dictionary as backing field
and to save the data of this dictionary for exported games I'm
using a property set with PROPERTY_USAGE_STORAGE that reads
and sets this backing field.